Prepare 4.1.5 with Debian load-surface and warning cleanup
Remove the standalone test translation unit from package builds and patch signedness/MSVC pragma warnings, while persisting these fixes in update_from_upstream.sh and adding tmp artifact ignore rules for clean CRAN checks.

+++ src/DgZ3StringRF.cpp
@@ -119,7 +119,7 @@ DgZ3StringRF::invQuantify (const DgHierNdxStringCoord& addIn) const
     // z3 codes
     std::string radStr1 = "";
     std::string radStr2 = "";
-    for (int i = 0; i < z3str.length(); i += 2) {
+    for (size_t i = 0; i < z3str.length(); i += 2) {
        std::string z3code = z3str.substr(i, 2);
        if (z3code == "00") {
           radStr1 += "0";
